requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-multipart>=0.0.6

# MCP SDK
git+https://github.com/modelcontextprotocol/python-sdk.git

# Image processing
pillow>=9.5.0

# HTTP client
requests>=2.28.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.2

# For development
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
